Resolving conflicts
