+++ examples/Example_opencv.cpp
