+++ src/CVObjectDetection.h
