+++ src/ClipProcessingJobs.h
