+++ src/effects/Tracker.cpp
@@ -28,7 +28,7 @@
  * along with OpenShot Library. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "../../include/effects/Tracker.h"
+#include "effects/Tracker.h"
 
 using namespace openshot;
