+++ src/protobuf_messages/compileProtobufMessages.sh
@@ -3,4 +3,4 @@ protoc -I=./ --cpp_out=./ trackerdata.proto
 protoc -I=./ --cpp_out=./ objdetectdata.proto
 
 mv *.cc ../
-mv *.h ../../include/
+mv *.h ../
