+++ tests/CVStabilizer_Tests.cpp
@@ -32,8 +32,8 @@
 #include "UnitTest++.h"
 // Prevent name clashes with juce::UnitTest
 #define DONT_SET_USING_JUCE_NAMESPACE 1
-#include "../include/OpenShot.h"
-#include "../include/ProcessingController.h"
+#include "OpenShot.h"
+//#include "ProcessingController.h"
 #include <QImage>
 
 using namespace openshot;
